gstrtpsession: Add parameter to control EOS events on RTCP pad
Currently, when a session is ended an a bye package is sent,
a EOS event is sent. If a bundle transport is used, this
causes other RTCP events, from other sources, not to reach its
destination because downstream element rejects more buffers. This
can be worse if no encryption is used. Once the EOS is sent,
all buffers are rejected RTCP and RTP.

Change-Id: I620aba608cf7954205b0d332c5a1bb90f377c17b