debian: Depend on exact version of gstreamer1.0-plugins-good
This makes the installation of kurento package to upgrade
gstreamer1.0-plugins-good, if it happened to be already installed in an
earlier (system default) version.

Otherwise, upon installation of gstreamer1.0-plugins-good-kurento, the
actual gstreamer1.0-plugins-good will not be upgraded.

Note that debian/control cannot be deleted: kurento-buildpackage expects
it to exist so build dependencies can be installed before attempting to
build the package itself.